// src/Types/Clinic.ts
export default interface ClinicType {
  id: string;
  first_name: string;
  mid_name: string;
  last_name: string;
  role_id: string;
  clinic_id: string;
  email: string;
  name: string;
  role: string;
  gender: string;
  short_name: string;
  contact_no: string;
  address: string;
  city: string;
  state: string;
  zip_code: string;
  address_2: string;
  phone: string;
  password?: string;
}